requirements.txt
praw>=7.7.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
pandas>=2.0.0
nltk>=3.8.0
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
